meta: reword the PyPI description, author-name diacritics, float64 doctest outputs
The project description now reads like the README rather than a keyword
list. The author and maintainer name carries its diacritics, matching
CITATION.cff. The vmap docstring examples printed float32 outputs in a
library that mandates float64.

+++ dftax/utils/vmap.py
@@ -124,7 +124,7 @@ def vmap(
     >>> vmap_square = vmap(square)
     >>> x = jnp.array([1.0, 2.0, 3.0])
     >>> vmap_square(x)  # doctest: +SKIP
-    Array([1., 4., 9.], dtype=float32)
+    Array([1., 4., 9.], dtype=float64)
 
     Vectorizing functions with multiple arguments and selective batch axes:
 
@@ -134,7 +134,7 @@ def vmap(
     >>> x = jnp.array([1.0, 2.0, 3.0])
     >>> y = jnp.array(10.0)
     >>> vmap_add(x, y)  # doctest: +SKIP
-    Array([11., 12., 13.], dtype=float32)
+    Array([11., 12., 13.], dtype=float64)
 
     Memory-efficient processing of large batches with chunking:
